Bug 1777952 - Namespace SurfaceTexture handles by the process they were allocated in. r=gfx-reviewers,geckoview-reviewers,aosmond,calu
When Surfaces/SurfaceTextures are allocated they are given a handle,
which is a monotonically increasing 32-bit integer. To render
Surfaces, we typically pass the Surface handle to the compositor,
which then looks it up in a map to find the corresponding
SurfaceTexture.

Following a GPU process restart, content may be left with stale
handles referencing SurfaceTextures which no longer exist. Once new
SurfaceTextures are allocated, these stale handles may reference new
SurfaceTextures with no relation to the old handle. This can lead to
rendering the wrong texture. Additionally, we may crash when
allocating "sync" SurfaceTextures, as the previous sync texture for a
certain handle may not have been released yet.

To fix this, this patch combines the existing handle with a new ID
uniquely identifying the process in which the SurfaceTexture was
allocated (or 0 for the parent process). We use a monotonically
increasing value rather than the pid to guard against the new GPU
process possibly having the same pid as the previous instance. We
combine these two 32-bit integers and use the resulting 64-bit integer
as the Surface handle.

Differential Revision: https://phabricator.services.mozilla.com/D150963

+++ mobile/android/geckoview/src/main/java/org/mozilla/gecko/gfx/SyncConfig.java
@@ -8,13 +8,13 @@ import android.os.Parcel;
 import android.os.Parcelable;
 
 /* package */ final class SyncConfig implements Parcelable {
-  final int sourceTextureHandle;
+  final long sourceTextureHandle;
   final GeckoSurface targetSurface;
   final int width;
   final int height;
 
   /* package */ SyncConfig(
-      final int sourceTextureHandle,
+      final long sourceTextureHandle,
       final GeckoSurface targetSurface,
       final int width,
       final int height) {
@@ -38,7 +38,7 @@ import android.os.Parcelable;
       };
 
   private SyncConfig(final Parcel parcel) {
-    sourceTextureHandle = parcel.readInt();
+    sourceTextureHandle = parcel.readLong();
     targetSurface = GeckoSurface.CREATOR.createFromParcel(parcel);
     width = parcel.readInt();
     height = parcel.readInt();
@@ -51,7 +51,7 @@ import android.os.Parcelable;
 
   @Override
   public void writeToParcel(final Parcel parcel, final int flags) {
-    parcel.writeInt(sourceTextureHandle);
+    parcel.writeLong(sourceTextureHandle);
     targetSurface.writeToParcel(parcel, flags);
     parcel.writeInt(width);
     parcel.writeInt(height);
